+++ src/components/SettingsDialog.tsx
@@ -387,6 +387,16 @@ const SettingsDialog: React.FC<SettingsDialogProps> = ({
                 updateSettings={updateSettings}
               />
             </SettingsSubPanel>
+            <SettingsSubPanel
+              label={i18n.get("Zoom indicator")}
+              value={getOnOff(settings.zoomEnabled)}
+            >
+              <ToggleSetting
+                propertyName={"zoomEnabled"}
+                settings={settings}
+                updateSettings={updateSettings}
+              />
+            </SettingsSubPanel>
             <SettingsSubPanel label={i18n.get("On dataset selection")}>
               <TextField
                 variant="standard"